Switch to new CSV format
