+++ scripts/spike_log_to_trace_csv.py
@@ -30,11 +30,7 @@ from lib import *
 RD_RE    = re.compile(r"(?P<pri>\d) 0x(?P<addr>[a-f0-9]+?) " \
                       "\((?P<bin>.*?)\) (?P<reg>[xf]\s*\d*?) 0x(?P<val>[a-f0-9]+)")
 CORE_RE  = re.compile(r"core.*0x(?P<addr>[a-f0-9]+?) \(0x(?P<bin>.*?)\) (?P<instr>.*?)$")
-INSTR_RE = re.compile(r"(?P<instr>[a-z\.0-9]+?)(\s+?)(?P<operand>.*)")
-GPR_RE   = re.compile(r"^[a-z][0-9a-z]$")
 ILLE_RE  = re.compile(r"trap_illegal_instruction")
-PC_RE    = re.compile(r"pc+")
-HEX_RE   = re.compile(r"^0x")
 
 LOGGER = logging.getLogger()
 
@@ -55,9 +51,6 @@ def process_spike_sim_log(spike_log, csv, full_trace = 0):
   cmd = ("sed -i '/ecall/q' %s" % spike_log)
   os.system(cmd)
 
-  gpr = {}
-  gpr["zero"] = 0
-
   with open(spike_log, "r") as f, open(csv, "w") as csv_fd:
     trace_csv = RiscvInstructionTraceCsv(csv_fd)
     trace_csv.start_new_trace()
@@ -69,9 +62,16 @@ def process_spike_sim_log(spike_log, csv, full_trace = 0):
         spike_instr = m.group("instr").replace("pc + ", "")
         spike_instr = spike_instr.replace("pc - ", "-")
         rv_instr_trace = RiscvInstructionTraceEntry()
+        rv_instr_trace.pc = m.group("addr")
         rv_instr_trace.instr_str = spike_instr
-        rv_instr_trace.addr = m.group("addr")
         rv_instr_trace.binary = m.group("bin")
+        if full_trace:
+          rv_instr_trace.instr = spike_instr.split(" ")[0]
+          rv_instr_trace.operand = spike_instr[len(rv_instr_trace.instr):]
+          rv_instr_trace.operand = rv_instr_trace.operand.replace(" ", "")
+          rv_instr_trace.instr, rv_instr_trace.operand = \
+            convert_pseudo_instr(rv_instr_trace.instr, rv_instr_trace.operand)
+          process_instr(rv_instr_trace)
         if spike_instr == "wfi":
           trace_csv.write_trace_entry(rv_instr_trace)
           continue
@@ -86,29 +86,31 @@ def process_spike_sim_log(spike_log, csv, full_trace = 0):
           m = RD_RE.search(nextline)
           if m:
             # Extract RD information
-            rv_instr_trace.rd = gpr_to_abi(m.group("reg").replace(" ",""))
-            rv_instr_trace.rd_val = m.group("val")
-            rv_instr_trace.privileged_mode = m.group("pri")
-            gpr[rv_instr_trace.rd] = rv_instr_trace.rd_val
+            rv_instr_trace.gpr.append(
+              gpr_to_abi(m.group("reg").replace(" ","")) + ":" + m.group("val"))
+            rv_instr_trace.mode = m.group("pri")
           else:
             # If full trace is not enabled, skip the entry that doesn't have
             # architectural state update.
             if not full_trace:
               continue
-          if full_trace:
-            s = INSTR_RE.search(spike_instr)
-            if s:
-              rv_instr_trace.instr = s.group("instr")
-              operand_str = s.group("operand").replace(" ", "")
-              operands = operand_str.split(",")
-              assign_operand(rv_instr_trace, operands, gpr)
-            else:
-              rv_instr_trace.instr = spike_instr
-        else:
-          rv_instr_trace.instr = spike_instr
         trace_csv.write_trace_entry(rv_instr_trace)
   logging.info("Processed instruction count : %d" % instr_cnt)
   logging.info("CSV saved to : %s" % csv)
+
+
+def process_instr(trace):
+  if trace.instr == "jal":
+    # Spike jal format jal rd, -0xf -> jal rd, -15
+    idx = trace.operand.rfind(",")
+    imm = trace.operand[idx+1:]
+    if imm[0] == "-":
+      imm = "-" + str(int(imm[1:], 16))
+    else:
+      imm = str(int(imm, 16))
+    trace.operand = trace.operand[0:idx+1] + imm
+  trace.operand = trace.operand.replace("(", ",")
+  trace.operand = trace.operand.replace(")", "")
 
 
 def main():
