+++ test/riscv_instr_test_pkg.sv
@@ -22,6 +22,5 @@ package riscv_instr_test_pkg;
   `include "riscv_instr_base_test.sv"
   `include "riscv_instr_test_lib.sv"
   `include "riscv_instr_cov_test.sv"
-  `include "exp_riscv_instr_cov_test.sv"
 
 endpackage
